+++ spec/api_stubs/stubs.rb
@@ -7,24 +7,23 @@ module FaaStRuby
       def initialize()
       end
 
-      def create_workspace_response(name = 'test-workspace', email = nil)
+      def create_workspace_response(name = 'test-workspace', email = nil, provider = nil)
         hash = {
           'object' => 'workspace',
           'name' => name,
-          'credentials' => {
-            'api_secret' => 'API_SECRET',
-            'api_key' => 'API_KEY'
-          }, 'created_at' => '2018-10-22 19:02:51 +0000',
+          'credentials' => credentials_example,
+          'created_at' => '2018-10-22 19:02:51 +0000',
           'errors' => [],
           'updated_at' => '2018-10-22 19:02:52 +0000'
         }
 
         hash['email'] = email if email
+        hash['provider'] = provider if provider
         hash
       end
 
-      def get_workspace_response(name = 'test-workspace', email = nil)
-        hash = create_workspace_response(name, email)
+      def get_workspace_response(name = 'test-workspace', email = nil, provider = nil)
+        hash = create_workspace_response(name, email, provider)
         hash.delete('created_at')
         hash['credentials'].delete('api_secret')
         hash
@@ -39,8 +38,15 @@ module FaaStRuby
         }
       end
 
-      def stub_create_workspace(name:, email: nil, status: 200)
-        stub_request(:post, "#{FAASTRUBY_HOST}/v2/workspaces").to_return(body: create_workspace_response(name, email).to_json, status: status)
+      def credentials_example
+        {
+          'api_secret' => 'API_SECRET',
+          'api_key' => 'API_KEY'
+        }
+      end
+
+      def stub_create_workspace(name:, email: nil, status: 200, provider: nil)
+        stub_request(:post, "#{FAASTRUBY_HOST}/v2/workspaces").to_return(body: create_workspace_response(name, email, provider).to_json, status: status)
       end
 
       def stub_destroy_workspace(name:, status: 200)
@@ -53,6 +59,10 @@ module FaaStRuby
 
       def stub_get_workspace(name:, status: 200)
         stub_request(:get, "#{FAASTRUBY_HOST}/v2/workspaces/#{name}").to_return(body: get_workspace_response(name).to_json, status: status)
+      end
+
+      def stub_refresh_credentials(name:, status: 200)
+        stub_request(:put, "#{FAASTRUBY_HOST}/v2/workspaces/#{name}/credentials").to_return(body: {name => credentials_example}.to_json, status: status)
       end
 
       def stub_run_function(method:, function_name:, workspace_name:, status: 200, query: nil)