+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.2.3 - Unreleased
+- Added #status_code to Workspace class to hold the API response code after a request
+- Added refresh_credentials endpoint
+- Fix the request headers
+
 ## 0.2.2 - Oct 20 2018
 ### New
 - Functions can be scheduled via faastruby.yml